requirements.txt
https://github.com/galperins4/python-client/archive/develop.zip
psycopg2-binary==2.7.4
flask==0.12.2
